Grid items to scale with Unity
items are to scale with Unity kitchen objects now, Grid is not to scale yet
