+++ frontend/src/components/floorPlan/grid.js
@@ -1,218 +1,109 @@
-// 2D GRID FOR THE FLOOR PLAN
-
-import React from 'react';
-import { useHistory } from "react-router-dom";
-import { makeStyles } from "@material-ui/core/styles";
-import { Box, Button, Typography } from "@material-ui/core";
-import styled from "styled-components";
-import _ from "lodash";
-import { Responsive, WidthProvider, GridLayout } from "react-grid-layout";
-import 'react-grid-layout/css/styles.css';
-import 'react-resizable/css/styles.css';
+import React, { useState } from "react";
+import { Responsive, WidthProvider } from "react-grid-layout";
+import "react-grid-layout/css/styles.css";
+import "react-resizable/css/styles.css";
+//import "./App.css";
 import './styles.css';
 
 const ResponsiveReactGridLayout = WidthProvider(Responsive);
 
-class ToolBoxItem extends React.Component {
-  render() {
-    return (
-      <div
-        className="toolbox__items__item"
-        onClick={this.props.onTakeItem.bind(undefined, this.props.item)}
+const ReactGridLayout = () => {
+  const [layouts, setLayouts] = useState(null);
+  const [widgetArray, setWidgetArray] = useState([
+    // { i: "widget1", x: 0, y: 0, w: 2, h: 2 },
+    // { i: "widget2", x: 2, y: 2, w: 2, h: 2 },
+    // { i: "widget3", x: 4, y: 4, w: 2, h: 2 },
+  ]);
+
+  const handleModify = (layouts, layout) => {
+    const tempArray = widgetArray;
+    setLayouts(layout);
+    layouts?.map((position) => {
+      tempArray[Number(position.i)].x = position.x;
+      tempArray[Number(position.i)].y = position.y;
+      tempArray[Number(position.i)].width = position.w;
+      tempArray[Number(position.i)].height = position.h;
+    });
+    console.log(tempArray)
+    setWidgetArray(tempArray);
+  };
+
+  const handleAdd = (name) => {
+    setWidgetArray([
+      ...widgetArray,
+      // { i: name + " " + (widgetArray.length + 1), x: 0, y: 0, w: 1, h: 1, type: name },
+      //previous w: 1, h: 1. Now these new dimensions are to scale with Unity
+      { i: name, x: 0, y: 0, w: 1.4, h: 0.7, type: name },
+    ]);
+  };
+
+  const handleDelete = (key) => {
+    const tempArray = widgetArray.slice();
+    const index = tempArray.indexOf(tempArray.find((data) => data.i === key));
+    tempArray.splice(index, 1);
+    setWidgetArray(tempArray);
+  };
+
+  return (
+    <div>
+      <button onClick={() => handleAdd("Table")}>Add Table</button>
+      <button onClick={() => handleAdd("Grill")}>Add Grill</button>
+      <button onClick={() => handleAdd("Fridge")}>Add Fridge</button>
+      <button onClick={() => handleAdd("Sink")}>Add Sink</button>
+      <button onClick={() => handleAdd("Microwave")}>Add Microwave</button>
+      <button onClick={() => handleAdd("Submission Station")}>Add Submission Station</button>
+
+      <ResponsiveReactGridLayout
+        onLayoutChange={handleModify}
+        verticalCompact={false}
+        layout={layouts}
+        breakpoints={{ lg: 1200, md: 996, sm: 768, xs: 480, xxs: 0 }}
+        preventCollision={true}
+        cols={{ lg: 8, md: 8, sm: 4, xs: 2, xxs: 2 }}
+        autoSize={true}
+        margin={{
+          lg: [20, 20],
+          md: [20, 20],
+          sm: [20, 20],
+          xs: [20, 20],
+          xxs: [20, 20],
+        }}
       >
-        {this.props.item.i}
-      </div>
-    );
-  }
-}
-class ToolBox extends React.Component {
-  render() {
-    return (
-      <div className="toolbox" style={{backgroundColor:"yellow"}}>
-        <span className="toolbox__title">Toolbox</span>
-        <div className="toolbox__items">
-          {this.props.items.map(item => (
-            <ToolBoxItem
-              key={item.i}
-              item={item}
-              onTakeItem={this.props.onTakeItem}
-            />
-          ))}
-        </div>
-      </div>
-    );
-  }
-}
-
-export default class ToolboxLayout extends React.Component {
-  static defaultProps = {
-    className: "layout",
-    rowHeight: 30,
-    onLayoutChange: function() {},
-    cols: { lg: 12, md: 10, sm: 6, xs: 4, xxs: 2 },
-    initialLayout: generateLayout()
-  };
-
-  state = {
-    currentBreakpoint: "lg",
-    compactType: "vertical",
-    mounted: false,
-    layouts: { lg: this.props.initialLayout },
-    toolbox: { lg: [] }
-  };
-
-  componentDidMount() {
-    this.setState({ mounted: true });
-  }
-
-  generateDOM() {
-    return _.map(this.state.layouts[this.state.currentBreakpoint], l => {
-      return (
-        <div key={l.i} className={l.static ? "static" : ""} style={{backgroundColor: "cyan"}}>
-          <div className="hide-button" onClick={this.onPutItem.bind(this, l)}>
-            &times;
-          </div>
-          {l.static ? (
-            <span
-              className="text"
-              title="This item is static and cannot be removed or resized."
+        {widgetArray?.map((widget, index) => {
+          return (
+            <div
+              className="reactGridItem"
+              style={{backgroundColor: "cyan"}}
+              key={index}
+              data-grid={{
+                x: widget?.x,
+                y: widget?.y,
+                w: widget?.w,
+                h: widget?.h,
+                i: widget.i,
+                //minW: 1, maxW: Infinity, minH: 1, maxH: Infinity
+                //These new dimensions are to scale with Unity
+                minW: 1.4,
+                maxW: 1.4, 
+                minH: 0.7,
+                maxH: 0.7,
+                isDraggable: true,
+                isResizable: true,
+              }}
             >
-              Static - {l.i}
-            </span>
-          ) : (
-            <span className="text">oven</span>
-          )}
-        </div>
-      );
-    });
-  }
+              <button
+                className="deleteButton"
+                onClick={() => handleDelete(widget.i)}
+              >
+                x
+              </button>
+              <div>{widget.i}</div>
+            </div>
+          );
+        })}
+      </ResponsiveReactGridLayout>
+    </div>
+  );
+};
 
-  onBreakpointChange = breakpoint => {
-    this.setState(prevState => ({
-      currentBreakpoint: breakpoint,
-      toolbox: {
-        ...prevState.toolbox,
-        [breakpoint]:
-          prevState.toolbox[breakpoint] ||
-          prevState.toolbox[prevState.currentBreakpoint] ||
-          []
-      }
-    }));
-  };
-
-  onCompactTypeChange = () => {
-    const { compactType: oldCompactType } = this.state;
-    const compactType =
-      oldCompactType === "horizontal"
-        ? "vertical"
-        : oldCompactType === "vertical"
-        ? null
-        : "horizontal";
-    this.setState({ compactType });
-  };
-
-  onTakeItem = item => {
-    this.setState(prevState => ({
-      toolbox: {
-        ...prevState.toolbox,
-        [prevState.currentBreakpoint]: prevState.toolbox[
-          prevState.currentBreakpoint
-        ].filter(({ i }) => i !== item.i)
-      },
-      layouts: {
-        ...prevState.layouts,
-        [prevState.currentBreakpoint]: [
-          ...prevState.layouts[prevState.currentBreakpoint],
-          item
-        ]
-      }
-    }));
-  };
-
-  onPutItem = item => {
-    this.setState(prevState => {
-      return {
-        toolbox: {
-          ...prevState.toolbox,
-          [prevState.currentBreakpoint]: [
-            ...(prevState.toolbox[prevState.currentBreakpoint] || []),
-            item
-          ]
-        },
-        layouts: {
-          ...prevState.layouts,
-          [prevState.currentBreakpoint]: prevState.layouts[
-            prevState.currentBreakpoint
-          ].filter(({ i }) => i !== item.i)
-        }
-      };
-    });
-  };
-
-  onLayoutChange = (layout, layouts) => {
-    this.props.onLayoutChange(layout, layouts);
-    this.setState({ layouts });
-  };
-
-  onNewLayout = () => {
-    this.setState({
-      layouts: { lg: generateLayout() }
-    });
-  };
-
-  render() {
-    return (
-      <div >
-        <div>
-          Current Breakpoint: {this.state.currentBreakpoint} (
-          {this.props.cols[this.state.currentBreakpoint]} columns)
-        </div>
-        <div>
-          Compaction type:{" "}
-          {_.capitalize(this.state.compactType) || "No Compaction"}
-        </div>
-        <button onClick={this.onNewLayout}>Generate New Layout</button>
-        <button onClick={this.onCompactTypeChange}>
-          Change Compaction Type
-        </button>
-
-        <ToolBox
-          items={this.state.toolbox[this.state.currentBreakpoint] || []}
-          onTakeItem={this.onTakeItem}
-        />
-
-        <ResponsiveReactGridLayout
-          {...this.props}
-          layouts={this.state.layouts}
-          onBreakpointChange={this.onBreakpointChange}
-          onLayoutChange={this.onLayoutChange}
-          // WidthProvider option
-          measureBeforeMount={false}
-          // I like to have it animate on mount. If you don't, delete `useCSSTransforms` (it's default `true`)
-          // and set `measureBeforeMount={true}`.
-          useCSSTransforms={this.state.mounted}
-          compactType={this.state.compactType}
-          preventCollision={!this.state.compactType}
-        >
-          {this.generateDOM()}
-        </ResponsiveReactGridLayout>
-      </div>
-    );
-  }
-}
-
-//generate the kitchen items one by one here somehow i think*****
-function generateLayout() {
-  return _.map(_.range(0, 5), function(item, i) {
-    var y = Math.ceil(Math.random() * 4) + 1;
-    return {
-      x: (_.random(0, 5) * 2) % 12,
-      y: Math.floor(i / 6) * y,
-      w: 2, //width of block
-      h: y, //height of block
-      i: i.toString(),
-      static: Math.random() < 0.05
-    };
-  });
-}
-    
+export default ReactGridLayout;
