+++ frontend/src/components/floorPlan/gridOld.js
@@ -1,0 +1,218 @@
+// old version 2D GRID FOR THE FLOOR PLAN
+
+import React from 'react';
+import { useHistory } from "react-router-dom";
+import { makeStyles } from "@material-ui/core/styles";
+import { Box, Button, Typography } from "@material-ui/core";
+import styled from "styled-components";
+import _ from "lodash";
+import { Responsive, WidthProvider, GridLayout } from "react-grid-layout";
+import 'react-grid-layout/css/styles.css';
+import 'react-resizable/css/styles.css';
+import './styles.css';
+
+const ResponsiveReactGridLayout = WidthProvider(Responsive);
+
+class ToolBoxItem extends React.Component {
+  render() {
+    return (
+      <div
+        className="toolbox__items__item"
+        onClick={this.props.onTakeItem.bind(undefined, this.props.item)}
+      >
+        {this.props.item.i}
+      </div>
+    );
+  }
+}
+class ToolBox extends React.Component {
+  render() {
+    return (
+      <div className="toolbox" style={{backgroundColor:"yellow"}}>
+        <span className="toolbox__title">Toolbox</span>
+        <div className="toolbox__items">
+          {this.props.items.map(item => (
+            <ToolBoxItem
+              key={item.i}
+              item={item}
+              onTakeItem={this.props.onTakeItem}
+            />
+          ))}
+        </div>
+      </div>
+    );
+  }
+}
+
+export default class ToolboxLayout extends React.Component {
+  static defaultProps = {
+    className: "layout",
+    rowHeight: 30,
+    onLayoutChange: function() {},
+    cols: { lg: 12, md: 10, sm: 6, xs: 4, xxs: 2 },
+    initialLayout: generateLayout()
+  };
+
+  state = {
+    currentBreakpoint: "lg",
+    compactType: "vertical",
+    mounted: false,
+    layouts: { lg: this.props.initialLayout },
+    toolbox: { lg: [] }
+  };
+
+  componentDidMount() {
+    this.setState({ mounted: true });
+  }
+
+  generateDOM() {
+    return _.map(this.state.layouts[this.state.currentBreakpoint], l => {
+      return (
+        <div key={l.i} className={l.static ? "static" : ""} style={{backgroundColor: "cyan"}}>
+          <div className="hide-button" onClick={this.onPutItem.bind(this, l)}>
+            &times;
+          </div>
+          {l.static ? (
+            <span
+              className="text"
+              title="This item is static and cannot be removed or resized."
+            >
+              Static - {l.i}
+            </span>
+          ) : (
+            <span className="text">oven</span>
+          )}
+        </div>
+      );
+    });
+  }
+
+  onBreakpointChange = breakpoint => {
+    this.setState(prevState => ({
+      currentBreakpoint: breakpoint,
+      toolbox: {
+        ...prevState.toolbox,
+        [breakpoint]:
+          prevState.toolbox[breakpoint] ||
+          prevState.toolbox[prevState.currentBreakpoint] ||
+          []
+      }
+    }));
+  };
+
+  onCompactTypeChange = () => {
+    const { compactType: oldCompactType } = this.state;
+    const compactType =
+      oldCompactType === "horizontal"
+        ? "vertical"
+        : oldCompactType === "vertical"
+        ? null
+        : "horizontal";
+    this.setState({ compactType });
+  };
+
+  onTakeItem = item => {
+    this.setState(prevState => ({
+      toolbox: {
+        ...prevState.toolbox,
+        [prevState.currentBreakpoint]: prevState.toolbox[
+          prevState.currentBreakpoint
+        ].filter(({ i }) => i !== item.i)
+      },
+      layouts: {
+        ...prevState.layouts,
+        [prevState.currentBreakpoint]: [
+          ...prevState.layouts[prevState.currentBreakpoint],
+          item
+        ]
+      }
+    }));
+  };
+
+  onPutItem = item => {
+    this.setState(prevState => {
+      return {
+        toolbox: {
+          ...prevState.toolbox,
+          [prevState.currentBreakpoint]: [
+            ...(prevState.toolbox[prevState.currentBreakpoint] || []),
+            item
+          ]
+        },
+        layouts: {
+          ...prevState.layouts,
+          [prevState.currentBreakpoint]: prevState.layouts[
+            prevState.currentBreakpoint
+          ].filter(({ i }) => i !== item.i)
+        }
+      };
+    });
+  };
+
+  onLayoutChange = (layout, layouts) => {
+    this.props.onLayoutChange(layout, layouts);
+    this.setState({ layouts });
+  };
+
+  onNewLayout = () => {
+    this.setState({
+      layouts: { lg: generateLayout() }
+    });
+  };
+
+  render() {
+    return (
+      <div >
+        <div>
+          Current Breakpoint: {this.state.currentBreakpoint} (
+          {this.props.cols[this.state.currentBreakpoint]} columns)
+        </div>
+        <div>
+          Compaction type:{" "}
+          {_.capitalize(this.state.compactType) || "No Compaction"}
+        </div>
+        <button onClick={this.onNewLayout}>Generate New Layout</button>
+        <button onClick={this.onCompactTypeChange}>
+          Change Compaction Type
+        </button>
+
+        <ToolBox
+          items={this.state.toolbox[this.state.currentBreakpoint] || []}
+          onTakeItem={this.onTakeItem}
+        />
+
+        <ResponsiveReactGridLayout
+          {...this.props}
+          layouts={this.state.layouts}
+          onBreakpointChange={this.onBreakpointChange}
+          onLayoutChange={this.onLayoutChange}
+          // WidthProvider option
+          measureBeforeMount={false}
+          // I like to have it animate on mount. If you don't, delete `useCSSTransforms` (it's default `true`)
+          // and set `measureBeforeMount={true}`.
+          useCSSTransforms={this.state.mounted}
+          compactType={this.state.compactType}
+          preventCollision={!this.state.compactType}
+        >
+          {this.generateDOM()}
+        </ResponsiveReactGridLayout>
+      </div>
+    );
+  }
+}
+
+//generate the kitchen items one by one here somehow i think*****
+function generateLayout() {
+  return _.map(_.range(0, 5), function(item, i) {
+    var y = Math.ceil(Math.random() * 4) + 1;
+    return {
+      x: (_.random(0, 5) * 2) % 12,
+      y: Math.floor(i / 6) * y,
+      w: 2, //width of block
+      h: y, //height of block
+      i: i.toString(),
+      static: Math.random() < 0.05
+    };
+  });
+}
+
